justicia-clara/requirements.txt
pydantic>=2
python-dotenv
requests
sentence-transformers
numpy
pypdf
streamlit
# OCR (optional; only if you’ll process scanned PDFs/images)
python-doctr[torch]
pdf2image
pillow
# torch and torchvision should be installed separately with:
# pip install torch torchvision --index-url https://download.pytorch.org/whl/cpu
